Significant change in admin view
Added container (AdminContainerPanel) to the admin view
(AdminMainPanel), so when we want to change panel for admin, we don't
have to change whole screen (logout & other buttons are always the
same). It works like an iframe in HTML. AdminMainPanel is static, and we
load different admin content into AdminContainerPanel as a cards. We
change only piece of view.